refactor(autonomi): move `get_closest_to_address` method to `network` mod from `quote`
