+++ autonomi/src/client/quote.rs
@@ -11,7 +11,7 @@ use crate::client::high_level::files::FILE_UPLOAD_BATCH_SIZE;
 use crate::client::utils::process_tasks_with_max_concurrency;
 use ant_evm::payment_vault::get_market_price;
 use ant_evm::{Amount, PaymentQuote, QuotePayment, QuotingMetrics};
-use ant_networking::{Addresses, Network, NetworkError};
+use ant_networking::{Network, NetworkError};
 use ant_protocol::{storage::ChunkAddress, NetworkAddress, CLOSE_GROUP_SIZE};
 use libp2p::PeerId;
 use std::collections::HashMap;
@@ -82,17 +82,6 @@ pub enum CostError {
 }
 
 impl Client {
-    /// Retrieve the closest peers to the given network address.
-    /// This function queries the network to find all peers in the close group nearest to the provided network address.
-    pub async fn get_closest_to_address(
-        &self,
-        network_address: &NetworkAddress,
-    ) -> Result<Vec<(PeerId, Addresses)>, NetworkError> {
-        self.network
-            .client_get_all_close_peers_in_range_or_close_group(network_address)
-            .await
-    }
-
     /// Get raw quotes from nodes.
     /// These quotes do not include actual record prices.
     /// You will likely want to use `get_store_quotes` instead.
